requirements.txt
langchain
langchain-community
langgraph
sqlalchemy
python-dotenv
pydantic
groq
jupyterlab
ipython
